+++ src/palm-react-native-ui-kit/screens/channels/EditChannelScreen/index.tsx
@@ -20,8 +20,7 @@ import React, { ReactElement } from 'react'
 import { useTranslation } from 'react-i18next'
 import { ScrollView, StyleSheet, TouchableOpacity, View } from 'react-native'
 import Icon from 'react-native-vector-icons/Ionicons'
-
-import TokenGating from './TokenGating'
+import TokenGating from '../TokenGating'
 
 const EditChannelScreen = (): ReactElement => {
   const { navigation, params } = useAppNavigation<Routes.EditChannel>()
@@ -205,7 +204,7 @@ const EditChannelScreen = (): ReactElement => {
       </Container>
       {showTokenGating && (
         <View style={{ position: 'absolute', width: '100%', height: '100%' }}>
-          <TokenGating useEditChannelReturn={useEditChannelReturn} />
+          <TokenGating useChannelReturn={useEditChannelReturn} />
         </View>
       )}
     </>
